Create autoposting.py
This Python script sends a photo with a caption to a specified chat ID using the aiogram library. The photo URL and caption text are extracted from a JSON file. If there is an error sending the message, the script will print the error message to the console. Once the message is sent, the script will close the dispatcher storage and bot session, and stop the bot.